Added handling for zero exchange rate and tax year info
In the `GetTaxReportDetails` method, code has been added to identify and store transactions with a zero `ExchangeRateValue` in the `invalidTransactions` list. A comment has been added to indicate future plans to augment these transactions with exchange rate data. Additionally, the `TaxYear` property is now set in the `TaxRecord` by calling the `GetAustralianTaxYear` method with the `TransactionDate`. This will provide tax year information for each taxable transaction.